backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.0

# Data & plotting
numpy
pandas
matplotlib